chore: extract rovodev client, update references
