+++ src/rovo-dev/client/responseParserInterfaces.ts
